speech-to-speech-gTTs/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
future==1.0.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai==1.55.2
openai-whisper @ git+https://github.com/openai/whisper.git@fc5ded7d9045c693692f13853857c3f8baea3a7b
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.11.0
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
